Finish homework - week 7 - Forms

+++ src/Controller/ProductController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Category;
 use App\Entity\Product;
+use App\Form\ProductType;
 use App\Repository\ProductRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -11,7 +12,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/product')]
+#[Route('/admin/product')]
 class ProductController extends AbstractController
 {
     private $entityManager;
@@ -26,59 +27,72 @@ class ProductController extends AbstractController
         $this->productRepository = $productRepository;
     }
 
-    #[Route('/', name: 'app_product')]
+    #[Route('/', name: 'app_admin_product')]
     public function index(): Response
     {
-        return $this->render('product/index.html.twig');
-    }
-
-    #[Route('/add', name: 'app_product_create')]
-    public function createProduct(): Response
-    {
-        $product = new Product();
-        $category = new Category();
-
-        $suffixCategory = rand(1,1000);
-        $category->setName('T-shirt ' . $suffixCategory);
-
-        $starPrice = rand(30,90);
-        $price = $starPrice + 90.99;
-
-        $product->setName('Koszulka')
-            ->setDescription('Biała z siateczką')
-            ->setPrice($price)
-            ->setActive(true)
-            ->setEan('12343212233');
-        
-        $product->setCategory($category);
-        
-        $this->entityManager->persist($category);
-        $this->entityManager->persist($product);
-        $this->entityManager->flush();
-
-        return $this->redirectToRoute('app_product');
-    }
-
-    #[Route('/list', name: 'app_product_list')]
-    public function list(Request $request): Response
-    {
-        $products = $this->productRepository->findAll();
-
-        $routeName = $request->attributes->get('_route');
+        /* $routeName = $request->attributes->get('_route');
         $routeParameters = $request->attributes->get('_route_params');
         // use this to get all the available attributes (not only routing ones):
         $allAttributes = $request->attributes->all();
 
         dump($routeName);
         dump($routeParameters);
-        dump($allAttributes);
+        dump($allAttributes); */
 
-        return $this->render('product/list.html.twig', [
-            'products' => $products,
+        $products = $this->productRepository->findAll();
+
+        return $this->render('product/index.html.twig', [
+            'products' => $products
         ]);
     }
 
-    #[Route('/show/{id}', name: 'app_product_show', requirements: ['id' => '\d+'])]
+    #[Route('/new', name: 'app_admin_product_new')]
+    public function new(Request $request): Response
+    {
+        $form = $this->createForm(ProductType::class);
+
+        $form->handleRequest($request);
+        if($form->isSubmitted() && $form->isValid())
+        {
+            $product = $form->getData();
+
+            $this->entityManager->persist($product);
+            $this->entityManager->flush();
+
+            $this->addFlash('success', 'Produkt został dodany');
+
+            return $this->redirectToRoute('app_admin_product');
+        }
+
+        return $this->render('product/edit.html.twig', [
+            'form' => $form->createView()
+        ]);
+    }
+
+    #[Route('/edit/{id}', name: 'app_admin_product_edit')]
+    public function edit(Product $product, Request $request): Response
+    {
+        $form = $this->createForm(ProductType::class, $product);
+
+        $form->handleRequest($request);
+        if($form->isSubmitted() && $form->isValid())
+        {
+            $product = $form->getData();
+
+            $this->entityManager->persist($product);
+            $this->entityManager->flush();
+
+            $this->addFlash('success', 'Produkt został zaktualizowany');
+
+            return $this->redirectToRoute('app_admin_product');
+        }
+
+        return $this->render('product/new.html.twig', [
+            'form' => $form->createView()
+        ]);
+    }
+
+    #[Route('/show/{id}', name: 'app_admin_product_show', requirements: ['id' => '\d+'])]
     public function show(Request $request, int $id): Response
     {
         $product = $this->productRepository->find($id);
@@ -88,18 +102,21 @@ class ProductController extends AbstractController
         ]);
     }
 
-    #[Route('/delete/{id<\d+>}', name: 'app_product_delete')]
+    #[Route('/delete/{id<\d+>}', name: 'app_admin_product_delete')]
     public function delete(Request $request, int $id): Response
     {
-        $product = $this->productRepository->find($id);
+        $product = $this->productRepository->findOneBy(['id' => $id]);
 
-        $this->entityManager->remove($product);
-        $this->entityManager->flush();
+        if($product) {
+            $this->entityManager->remove($product);
+            $this->entityManager->flush();
+        }
 
-        return $this->redirectToRoute('app_product_list');
+
+        return $this->redirectToRoute('app_admin_product');
     }
 
-    #[Route('/sales', name: 'app_product_sales')]
+    #[Route('/sales', name: 'app_admin_product_sales')]
     public function sales(Request $request): Response
     {
         $products = $this->productRepository->findAllLowerThanPrice(100);
@@ -109,7 +126,7 @@ class ProductController extends AbstractController
         ]);
     }
 
-    #[Route('/greater-qb', name: 'app_product_greater_qb')]
+    #[Route('/greater-qb', name: 'app_admin_product_greater_qb')]
     public function greaterQb(Request $request): Response
     {
         $products = $this->productRepository->findAllGreaterThanPrice(100, true);
